refactor(container): extract Binding and resolution helpers into focused modules
Split the 1,189-line container.py into three focused modules so each has a
single responsibility and binding lifecycle management can be read and tested
independently from container orchestration:

- inversipy/binding.py: `Binding` class, `ParameterDependency`, the cached
  `analyze_parameters()`, `_injectable_to_param_deps()`, `_extract_wrapper_type()`,
  `_extract_optional_type()`, and the `_MissingType` sentinel.
- inversipy/_wrappers.py: Factory/Lazy `_make_wrapper()` and
  `_make_wrapper_async()` plus the small `_format_dependency()` helper.
- inversipy/container.py: trimmed to only the `Container` class; imports the
  extracted helpers.

Pure internal reorganization — no public API changes. The single direct import
of `Binding` in tests/test_container.py is updated to its new location.

Closes #59
